requirements.txt
hunspell
pyocr
tesserocr
tesseract
PyPDF2
wand
Flask==0.12.2
gunicorn==19.1.1
pytesseract
pillow
pyyaml
dateparser
unidecode
pdfminer.six
python-box
sklearn